+++ client/src/app/transactions/category-input.tsx
@@ -13,7 +13,7 @@ import { cn } from '@/lib/utils';
 import { categories, type Category } from '@/types/transaction';
 import { Check, ChevronsUpDown } from 'lucide-react';
 import React from 'react';
-import CommandSubcategory from '../command-subcategory';
+import CommandSubcategory from '../dashboard/command-subcategory';
 
 interface CategoryInputProps {
   initialValue: string;
