+++ client/src/app/transactions/columns.tsx
@@ -2,7 +2,7 @@ import { type RowData, type ColumnDef } from '@tanstack/react-table';
 import { type Transaction } from '@/types/transaction';
 import DataTableHeader from '@/components/data-table-header';
 import EditableCell from './editable-cell';
-import LoadingCell from '../../../components/loading-cell';
+import LoadingCell from '../../components/loading-cell';
 // import AddTransactionButton from './add-transaction-button';
 
 declare module '@tanstack/react-table' {
